requirements.txt
Django==1.10.4
simplejson==3.10.0
djangorestframework==3.5.3
